refactor: App component and extract UI areas into separate components
