+++ src/components/index.tsx
@@ -1,0 +1,4 @@
+export { default as ParametersArea } from './ParametersArea'
+export { default as RunButtonArea } from './RunButtonArea'
+export { default as ErrorMessageArea } from './ErrorMessageArea'
+export { default as OutputArea } from './OutputArea'
